1. 基本完成管理员的页面 2. 基本完成接口调整和验证 TODO: 1. 用户页面和管理员的添加、详情页面 2. 接口验证

+++ src/main/java/com/example/entity/FinancialInstitution.java
@@ -17,6 +17,6 @@ import java.util.Map;
 @TableName(autoResultMap = true)
 public class FinancialInstitution extends Company {
     @TableField(typeHandler = JacksonTypeHandler.class)
-    private Map<String, BigDecimal> loanInterestRate;
+    private Map<String, Double> loanInterestRate;
     private String enterpriseIndustry;
 }
